# shared-contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Property Management System",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "description": "Authenticate user and return JWT tokens (FR-USER-01).\n\nRequest body\n------------\n.. code-block:: json\n\n    {\"email\": \"...\", \"password\": \"...\"}\n\nResponses\n---------\n200\n    ``{\"data\": {\"access_token\": \"...\", \"refresh_token\": \"...\", \"user\": {...}}}``\n401\n    ``{\"error\": {\"code\": \"AUTH-001\", \"message\": \"Invalid email or password\"}}``\n403\n    ``{\"error\": {\"code\": \"AUTH-002\", \"message\": \"Account is not active\"}}``",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Login Api V1 Auth Login Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "description": "Accept an invitation and create a new user account (FR-USER-02).\n\nRequest body\n------------\n.. code-block:: json\n\n    {\"invite_token\": \"...\", \"full_name\": \"...\", \"password\": \"...\", \"phone\": \"...\"}\n\nResponses\n---------\n201\n    ``{\"data\": {\"id\": \"...\", \"email\": \"...\", \"full_name\": \"...\", \"property_scopes\": [], \"is_active\": true}}``\n400\n    ``{\"error\": {\"code\": \"VAL-400\", \"message\": \"...\"}}`` (validation error)\n409\n    ``{\"error\": {\"code\": \"AUTH-004\", \"message\": \"Email already in use\"}}``",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Register Api V1 Auth Register Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/invite": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Invite",
        "description": "Send an invitation link to a new user (FR-USER-02).\n\nRequires the current user to have sufficient property scope\n(validated in the service layer).\n\nRequest body\n------------\n.. code-block:: json\n\n    {\"email\": \"...\", \"property_id\": \"...\"}\n\nResponses\n---------\n201\n    ``{\"data\": {\"invite_link\": \"...\"}}``\n403\n    ``{\"error\": {\"code\": \"AUTH-005\", \"message\": \"Insufficient property scope\"}}``\n409\n    ``{\"error\": {\"code\": \"AUTH-004\", \"message\": \"Email already in use\"}}``",
        "operationId": "invite_api_v1_auth_invite_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Invite Api V1 Auth Invite Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh",
        "description": "Issue a new access token from a valid refresh token (FR-USER-03).\n\nRequest body\n------------\n.. code-block:: json\n\n    {\"refresh_token\": \"...\"}\n\nResponses\n---------\n200\n    ``{\"data\": {\"access_token\": \"...\"}}``\n401\n    ``{\"error\": {\"code\": \"AUTH-007\", \"message\": \"Invalid or expired refresh token\"}}``",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Payload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Refresh Api V1 Auth Refresh Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get Me",
        "description": "Return the authenticated user's profile (FR-USER-01).\n\nRequires a valid Bearer token in the ``Authorization`` header.\n\nResponses\n---------\n200\n    ``{\"data\": {\"id\": \"...\", \"email\": \"...\", \"full_name\": \"...\", \"property_scopes\": [], \"is_active\": true}}``\n401\n    ``{\"error\": {\"code\": \"AUTH-009\", \"message\": \"Invalid or expired access token\"}}``",
        "operationId": "get_me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Me Api V1 Auth Me Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/audit-logs": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "View audit logs (paginated)",
        "description": "Returns paginated audit logs for a property. Owner role required.",
        "operationId": "get_audit_logs_api_v1_admin_audit_logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "property_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Filter by property",
              "title": "Property Id"
            },
            "description": "Filter by property"
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by action type",
              "title": "Action"
            },
            "description": "Filter by action type"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Items per page",
              "default": 50,
              "title": "Limit"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/config": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "View system configuration",
        "description": "Returns system settings with secrets masked. Owner role required.",
        "operationId": "get_system_config_api_v1_admin_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemConfigListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "AuditLogListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/AuditLogResponse"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "AuditLogListResponse"
      },
      "AuditLogResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "resource_type": {
            "type": "string",
            "title": "Resource Type"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "property_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Property Id"
          },
          "metadata": {
            "title": "Metadata"
          },
          "ip_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Address"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "action",
          "resource_type",
          "timestamp"
        ],
        "title": "AuditLogResponse"
      },
      "AuthRequest": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 255,
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "AuthRequest",
        "description": "Payload for ``POST /api/v1/auth/login`` (SDD \u00a73.3).\n\nValidates that the password meets minimum length and complexity\nrequirements before reaching the service layer."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InviteRequest": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 255,
            "format": "email",
            "title": "Email"
          },
          "property_id": {
            "type": "string",
            "format": "uuid",
            "title": "Property Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "property_id"
        ],
        "title": "InviteRequest",
        "description": "Payload for ``POST /api/v1/auth/invite`` (SDD \u00a73.3).\n\nAn admin sends an invitation for a specific email address and property.\n\n.. note::\n   Not strict so that string-encoded UUIDs from JSON deserialisation\n   are automatically coerced to ``uuid.UUID``."
      },
...
      "RegisterRequest": {
        "properties": {
          "invite_token": {
            "type": "string",
            "title": "Invite Token"
          },
          "full_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 2,
            "title": "Full Name"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password"
          },
          "phone": {
            "type": "string",
            "maxLength": 15,
            "minLength": 10,
            "title": "Phone"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "invite_token",
          "full_name",
          "password",
          "phone"
        ],
        "title": "RegisterRequest",
        "description": "Payload for ``POST /api/v1/auth/register`` (SDD \u00a73.3).\n\nCarries the invite token alongside the new user's profile data."
      },
...
      "SystemConfigListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/SystemConfigResponse"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "type": "null",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SystemConfigListResponse"
      },
      "SystemConfigResponse": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "masked": {
            "type": "boolean",
            "title": "Masked",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "key",
          "value"
        ],
        "title": "SystemConfigResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}